Package Info route
Nested inside the `package-list` route. Takes a package name as a slug and passes it to the model hook for `pkg-info` route.

The route controller makes a call to the cranDB API and pulls down details for the specified package as a JSON payload.

Template renders relevant details in the `package-list` nested outlet