Created `package-list` component
Moved the list of packages to the `package-list` component and added a `filteredPackages` computed property to the controller. This allows us to search for packages and automatically filter the list.

Working on an event handler so that you can press enter and automatically transition to the route for the top package in the list