Added route transitions to `package-list` search
Added an event handler in `package-list` component` for detecting when the user presses `Enter` in the search box.

`keyPress` detects the `Enter` key, grabs the first element from thepackage list, and passes it into the `onEnter` closure action.

Added a `packages` controller with a `showPkgInfo` action that takes a package ID as argument and triggers a route transition to the correct `pkg-info` route.

The `packages` template now passes the controller's `showPkgInfo` function as a closure action to the `package-list` component via the `onEnter` attribute.