Add search button functionality
Refactored the top-level action name to `onSearch`.

Extracted the search functionality to the `showTopResult` method on `package-list` and then linked the keypress event AND the `search` action to the method (so we're not repeating code like before).

The search button in `package-list` template now has a `search` action defined on it.