Basic shopping cart behavior for adding packages
The `cran-search` route has been expanded to include model components as well as the `package-cart` service that ties them all together.

`package-cart-item` component provides UI and behavior for items that are being stored by the `package-cart` service. In the future, it will also provide an interface into the services methods for adding/removing packages.