Implement `addPackage` behavior in cart service
The `package-cart` service now has a `getPkgInfo` method which takes a package from the cart, pulls the package information from CRAN, and uses it to create a new instance of the `Package` model.

The `package-cart-item` component now has an `addPackage` action that calls `getPkgInfo` from the service and then calls `Model.save()` on the result, persisting it to the Firebase backend.