Fetch existing sessions in the `application` route
Use the `beforeModel` hook to call EmberFire's `session.fetch()` method to grab an existing session (if it exists).

Fetching the existing session allows the app to recognize authenticated users even after page reloads.

We `return` the `fetch` method in order to force a synchronous fetch. This is useful because it makes the existing session available further down the route chain, allowing us to check if the user is authenticated and redirect accordingly.