Added a shell for login/signup action on the route
The `auth-form` component now has a `on-submit` argument that you can call when instantiating it.

Inside the component, the `login` and `signUp` actions will trigger the `on-submit` action with `Component.sendAction`.

In the `login` template, the `on-submit` action is mapped to the routes's `loginOrSignUp` action, which accepts an `isLogin` boolean and will dispatch a login or signUp method accordingly.