Add login and signup functionality
The `auth-form` component now grabs the form data, creates an object from it, and passes it to the `on-submit` action handler, where it is then sent to the `login` route.

On the `login` route, we implement both the `login` and `signUp` methods, each of which take a `userInfo` object as an argument and which authenticate and/or create the user based on the data received.

TO DO: Need to implement validations on signUp