Add redirects for non-authenticated users
`add-package` and `package-search` routes now redirect to `packages` if the user is not authenticated.

The nav links are already disabled for these routes if the user is not logged in, however, the redirection is an added layer of security since the user could theoretically manually type the other URLs in.