Refactor the login and signup to controller
Moved all of the auth functions to a `login` controller instead of being on the route. Having the functionality on the route was making it difficult to have the UI react to failed logins, so now the controller lets us track login state and reflect failed logins with an alert box.