Add form validation using `ember-validations`
The `auth-form` component now has built-in form validations for each field using the `ember-validations` addon combined with built-in validation functionality from `ember-bootstrap`'s form components.

The top-level `auth-form` component extends `EmberValidations` and implements the form properties and validations inside the component.

These properties are then passed into the `model` property of the `bs-form` components, which automatically triggers the bootstrap validation markup and subsequent behaviors.

One of the added features is that the `bs-form` component automatically stifles the `submit` action if any of the validations are failing, so we don't have to worry about preventing faulty form submissions to the server.