Added the ability to remove packages
Removing packages requires a two-step process whereby you delete and confirm, with the option to cancel the delete.

These behaviors are implemented in the `packages` controller and sent as closure actions through the `packages` route down to each individual list item.

Consequently, the the `packages` template has been refactored into a series of nested components. Whereas before we just had the top-level `package-list` component, the UI and behavior for individual item in the list has also been moved into its own `package-list-item` component.

`package-list-item` is where the UI for delete, confirm, and undo methods are implemented. The component receives the controller methods via closure action passed from the top-level and passes the package id back up to the controller methods