Refactor `package-cart` service to catch errors
Slight change to to the `pushToServer` method on `package-cart` service. The last link of the promise chain now has an error function to catch any errors returned from the backend.

Unfortunately, at this point it simply prevents the cart from removing that particular package, but doesn't prevent the `Model.save()` method from occuring, so the package will still show up in the local store until there's a refresh.

This is obviously less than ideal, as it makes it appear that the save occurred without error if you happen to navigate back to `packages` without checking the cart.

The solution to this is likely some combination of providing better success and failure feedback on the `search-packages` route itself and (possibly?) rolling back the save operation so that the package list doesn't falsely reflect the addition of th new package.